backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=1.0.0 